requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
rich==13.9.4
requests==2.32.3

# Dev / qualidade
pytest==8.3.3
ruff==0.7.4
